+++ src/validate/checks/index.js
@@ -4,4 +4,5 @@ module.exports = [
   require('./overlay-shape'),
   require('./overlay-targets'),
   require('./required-locales'),
+  require('./manifest-products'),
 ];